Move calculateMonthlyData() into MonthlyEmissionsCharts component
